+++ DataInterface/IDecisionRepository.cs
@@ -4,6 +4,6 @@ namespace DataInterface
 {
     public interface IDecisionRepository : IFormRepository<Decisions>
     {
-        Decisions GetByChildId(int childId);
+        Decisions GetByChildId(long childId);
     }
 }
